fix typo in project name folder
