+++ src/main/java/com/specialteam/coffeeshop/CoffeshopApplication.java
@@ -1,0 +1,12 @@
+package com.specialteam.coffeeshop;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CoffeshopApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CoffeshopApplication.class, args);
+	}
+}
