+++ src/main/java/com/specialteam/coffeeshop/config/WebSecurityConfig.java
@@ -1,0 +1,60 @@
+package com.specialteam.coffeeshop.config;
+
+import com.specialteam.coffeeshop.user.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+public class WebSecurityConfig {
+    public String encode(CharSequence rawPassword) {
+        return passwordEncoder.encode(rawPassword);
+    }
+
+    public boolean matches(CharSequence rawPassword, String encodedPassword) {
+        return passwordEncoder.matches(rawPassword, encodedPassword);
+    }
+
+    public boolean upgradeEncoding(String encodedPassword) {
+        return passwordEncoder.upgradeEncoding(encodedPassword);
+    }
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    private UserService userService;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        return http.csrf()
+                .disable()
+                .authenticationProvider(daoAuthenticationProvider())
+                .authorizeHttpRequests(
+                        (request) -> {
+                            request.requestMatchers("/api/v1/auth/register").permitAll();
+                            request.requestMatchers(HttpMethod.GET, "/api/v1/**").permitAll();
+                            request.requestMatchers(HttpMethod.POST, "/api/v1/**").hasRole("ROLES_USER");
+                        })
+                .authorizeHttpRequests()
+                .anyRequest()
+                .authenticated()
+                .and()
+                .formLogin(Customizer.withDefaults())
+                .httpBasic(Customizer.withDefaults())
+                .build();
+    }
+
+    @Bean
+    public DaoAuthenticationProvider daoAuthenticationProvider() {
+        DaoAuthenticationProvider provider = new DaoAuthenticationProvider(passwordEncoder);
+        provider.setUserDetailsService(userService);
+        return provider;
+    }
+}
