+++ src/main/java/com/specialteam/coffeeshop/product/dto/ProductRequestDto.java
@@ -1,0 +1,11 @@
+package com.specialteam.coffeeshop.product.dto;
+
+import lombok.Data;
+
+/** Product Request Body */
+@Data
+public class ProductRequestDto {
+    private String name;
+    private String description;
+    private Double price;
+}
