+++ src/main/java/com/specialteam/coffeeshop/product/dto/ProductResponseDto.java
@@ -1,0 +1,19 @@
+package com.specialteam.coffeeshop.product.dto;
+
+import java.time.Instant;
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+@Data
+public class ProductResponseDto<T> {
+    private Instant timestamp;
+    private T results;
+
+    public static <T> ResponseEntity<ProductResponseDto<T>> response(T results, HttpStatus status) {
+        ProductResponseDto<T> response = new ProductResponseDto<>();
+        response.setTimestamp(Instant.now());
+        response.setResults(results);
+        return new ResponseEntity<ProductResponseDto<T>>(response, status);
+    }
+}
