+++ src/main/java/com/specialteam/coffeeshop/user/model/User.java
@@ -1,0 +1,54 @@
+package com.specialteam.coffeeshop.user.model;
+
+import java.util.Collection;
+import java.util.List;
+import lombok.Data;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Document("users")
+@Data
+public class User implements UserDetails {
+    @Id
+    private String uid;
+
+    private String fullname;
+
+    private String username;
+
+    private String password;
+
+    private List<String> roles;
+
+    private Boolean isAccountNonExpired;
+
+    private Boolean isAccountNonLocked;
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return this.roles.stream().map(role -> new SimpleGrantedAuthority(role)).toList();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return isAccountNonExpired;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return isAccountNonLocked;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
