--- src/main/java/com/specialteam/coffeshop/CoffeshopApplication.java
@@ -1,12 +1,0 @@
-package com.specialteam.coffeshop;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoffeshopApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoffeshopApplication.class, args);
-	}
-}
