--- src/main/java/com/specialteam/coffeshop/product/dto/ProductRequestDto.java
@@ -1,13 +1,0 @@
-package com.specialteam.coffeshop.product.dto;
-
-import lombok.Data;
-
-/**
- * Product Request Body
- */
-@Data
-public class ProductRequestDto {
-    private String name;
-    private String description;
-    private Double price;
-}
