--- src/main/java/com/specialteam/coffeshop/user/model/User.java
@@ -1,56 +1,0 @@
-package com.specialteam.coffeshop.user.model;
-
-import java.util.Collection;
-import java.util.List;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import lombok.Data;
-
-@Document("users")
-@Data
-public class User implements UserDetails {
-    @Id
-    private String uid;
-
-    private String fullname;
-
-    private String username;
-
-    private String password;
-
-    private List<String> roles;
-
-    private Boolean isAccountNonExpired;
-
-    private Boolean isAccountNonLocked;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return this.roles.stream().map(role -> new SimpleGrantedAuthority(role)).toList();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return isAccountNonExpired;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return isAccountNonLocked;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
