--- src/main/java/com/specialteam/coffeshop/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.specialteam.coffeshop.user.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.specialteam.coffeshop.user.model.User;
-
-public interface UserRepository extends MongoRepository<User, String> {
-    Optional<User> findByUsername(String username);
-}
