--- src/test/java/com/specialteam/coffeshop/product/service/ProductServiceTest.java
@@ -1,16 +1,0 @@
-package com.specialteam.coffeshop.product.service;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@ExtendWith(MockitoExtension.class)
-public class ProductServiceTest {
-
-    @Test
-    void whenCalledGetProducts_shouldReturnValidSize() {
-
-    }
-}
